requirements.txt
adafruit-circuitpython-dht
adafruit-circuitpython-busdevice
adafruit_dht
RPi.GPIO